ABTesting small change
User rerolls if they are gonna participate to the test or not if the percentage received from the server is different than the previously received one

+++ shared/src/commonMain/kotlin/ai/lerna/multiplatform/service/dto/LernaAppConfig.kt
@@ -12,5 +12,5 @@ class LernaAppConfig {
 
 	var logSensorData: Boolean = true
 
-	var abTest: Double = 0.0
+	var abTest: Float = 0.0f
 }
